move ALIQUOT_FILTER_MODE to constants.ts
