+++ packages/components/src/internal/components/samples/SampleAssayDetail.spec.tsx
@@ -17,7 +17,8 @@ import { mountWithAppServerContext, mountWithServerContext, waitForLifecycle } f
 import { TEST_USER_AUTHOR, TEST_USER_READER } from '../../userFixtures';
 import { getTestAPIWrapper } from '../../APIWrapper';
 
-import { ALIQUOT_FILTER_MODE, SampleAliquotViewSelector } from './SampleAliquotViewSelector';
+import { SampleAliquotViewSelector } from './SampleAliquotViewSelector';
+import { ALIQUOT_FILTER_MODE } from './constants';
 import {
     AssayResultPanel,
     getSampleAssayDetailEmptyText,
